tests: fix broken tests

+++ source/memory/slowMemory.cpp
@@ -1,6 +1,6 @@
 #include "slowMemory.hpp"
 
-SlowMemory::SlowMemory(const uint16_t size)
+SlowMemory::SlowMemory(const uint32_t size)
 {
 	this->_data = std::vector<uint8_t>(size, 0);
 }
